Update KNOWN-ISSUES.
git-svn-id: https://repo.eecs.berkeley.edu/svn/users/ksen/projects/corvette-instrumentor@6368 1c3fd406-9068-4342-8353-511fcdf43622

+++ src/IValue.cpp
@@ -191,7 +191,7 @@ string IValue::toString() {
       s << "[FLP64: " << (double) value.as_flp << "]";
       break;
     case FLP80X86_KIND:
-      s << "[FLP80X86: " << (long double) value.as_flp << "]";
+      s << "[FLP80X86: " << (double) value.as_flp << "]";
       break;
     case VOID_KIND:
       s << "[VOID]";
